requirements.txt
# pip install -r requirements.txt
tqdm
h5py
numpy
opencv-python
pytorch==1.2.0
torchvision==0.4.0
tensorboard
scipy
opencv-python
matplotlib

